{"schema":"decodex/commit/1","summary":"Split core memory API routes","authority":"manual"}
